Integrating AMI sharing tests within packer basic example tests
